+++ src/components/ValutesExchange.tsx
@@ -1,12 +1,11 @@
-import React, { FC } from 'react'
+import { FC } from 'react'
 import { IExchange } from '../types/types'
 
 interface ValutesEchangeProps {
   exchangeData: IExchange[],
-  handleChange: (arg: string) => void,
 }
 
-const ValutesExchange:FC<ValutesEchangeProps> = ({exchangeData, handleChange}) => {
+const ValutesExchange:FC<ValutesEchangeProps> = ({exchangeData}) => {
   return (
     <div className='valutes__content'>
       <ul className='valutes__content-list'>
@@ -15,7 +14,7 @@ const ValutesExchange:FC<ValutesEchangeProps> = ({exchangeData, handleChange}) =
           <div className='valutes__content-item-value valutes__content-item_rate'>Rate</div>
         </li>
         {exchangeData.map(data => 
-          <li key={data.country} onClick={() => handleChange(data.country)} className='valutes__content-item'>
+          <li key={data.country} className='valutes__content-item'>
             <div className='valutes__content-item-value valutes__content-item_code'>{data.country}</div>
             <div className='valutes__content-item-value valutes__content-item_rate'>{data.val}</div>
           </li>
